requirements.txt
absl-py==0.10.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Deprecated==1.2.10
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
idna==2.10
iniconfig==1.1.1
Keras-Preprocessing==1.1.2
Markdown==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGithub==1.53
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
